Split random methods from math.hpp into separate random.hpp
Also add more tests for the random functions
