+++ FieldOpt/Optimization/optimizers/PSO.cpp
@@ -20,6 +20,7 @@
 
 #include "PSO.h"
 #include "Utilities/math.hpp"
+#include "Utilities/random.hpp"
 #include "Utilities/stringhelpers.hpp"
 #include <math.h>
