+++ FieldOpt/Utilities/tests/test_math.cpp
@@ -76,74 +76,5 @@ TEST_F(MathTest, Range) {
     EXPECT_FLOAT_EQ(0.9, ran3[ran3.size()-1]);
 }
 
-TEST_F(MathTest, RandomInts) {
-    auto gen = get_random_generator();
-
-    // random single int
-    int r1 = random_integer(gen, 0, 10);
-    int r2 = random_integer(gen, 0, 10);
-//    EXPECT_NE(r1, r2);
-    EXPECT_GE(r1, 0);
-    EXPECT_GE(r2, 0);
-    EXPECT_LE(r1, 10);
-    EXPECT_LE(r2, 10);
-
-    // random vector
-    std::vector<int> rands1 = random_integers(gen, 0, 10000, 10);
-    std::vector<int> rands2 = random_integers(gen, 0, 10000, 10);
-    EXPECT_EQ(10, rands1.size());
-
-    // Check that all numbers are different
-    for (int j = 0; j < rands1.size() - 1; ++j) {
-        EXPECT_NE(rands1[j], rands1[j+1]);
-    }
-
-    // Check that the two sets of rands are different
-    for (int i = 0; i < rands1.size(); ++i) {
-        EXPECT_NE(rands1[i], rands2[i]);
-    }
-}
-
-TEST_F(MathTest, RandomFloats) {
-    auto gen = get_random_generator();
-    std::vector<float> rands1 = random_floats(gen, 10);
-    std::vector<float> rands2 = random_floats(gen, 10);
-    EXPECT_EQ(10, rands1.size());
-
-    // Check that all numbers are different
-    for (int i = 0; i < rands1.size() - 1; ++i) {
-        EXPECT_NE(rands1[i], rands1[i+1]);
-    }
-
-    // Check that the two sets of rands are different
-    for (int j = 0; j < rands1.size(); ++j) {
-        EXPECT_NE(rands1[j], rands2[j]);
-    }
-}
-
-TEST_F(MathTest, SymmetricBernoulli) {
-	auto gen = get_random_generator(1);
-	auto numbers = random_symmetric_bernoulli(gen, 15);
-	EXPECT_EQ(15, numbers.size());
-	for (int number : numbers) {
-		EXPECT_TRUE(number == 1 || number == -1);
-	}
-}
-
-TEST_F(MathTest, SeededRandoms) {
-    auto gen11 = get_random_generator(1);
-    auto gen12 = get_random_generator(2);
-    auto gen21 = get_random_generator(1);
-    auto gen22 = get_random_generator(2);
-
-    double rand11 = random_double(gen11);
-    double rand12 = random_double(gen12);
-    double rand21 = random_double(gen21);
-    double rand22 = random_double(gen22);
-    EXPECT_EQ(rand11, rand21);
-    EXPECT_EQ(rand12, rand22);
-    EXPECT_NE(rand11, rand12);
-    EXPECT_NE(rand21, rand22);
-}
 
 }
