+++ FieldOpt/Utilities/verbosity.h
@@ -33,7 +33,7 @@
 
 //                    Module
 #define VERB_MOD 1 // Model
-#define VERB_OPT 2 // Optimization
+#define VERB_OPT 1 // Optimization
 #define VERB_WIC 1 // WellIndexCalculation
 #define VERB_SIM 1 // Simulation
 #define VERB_RUN 1 // Runner
